+++ cx-portal/src/components/shared/basic/ReleaseProcess/AppMarketCard/index.tsx
@@ -64,6 +64,7 @@ import CommonConnectorFormInputField from '../CommonFiles/CommonConnectorFormInp
 import ConnectorFormInputFieldShortAndLongDescription from '../CommonFiles/ConnectorFormInputFieldShortAndLongDescription'
 import ConnectorFormInputFieldImage from '../CommonFiles/ConnectorFormInputFieldImage'
 import ReleaseStepHeader from '../CommonFiles/ReleaseStepHeader'
+import { fetchImageWithToken } from 'services/ImageService'
 
 type FormDataType = {
   title: string