+++ CHANGELOG.md
@@ -23,6 +23,8 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * Fetch documents from backend and download
    * Update style multiple order section
    * Add short description to card
+AppReleaseProcess
+   * Conformity Document Type - Add Dropzone & Document handling
 
 ## 1.0.0-RC10